+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Nothing new
 - [#6](https://github.com/Nyphet/jsdoc-generator/issues/6) Vue.js files support (`.vue`).
 - [#9](https://github.com/Nyphet/jsdoc-generator/issues/9) Implementation for `generateJsdocFiles` and renamed it into `generateJsdocWorkspace`.
 - [#10](https://github.com/Nyphet/jsdoc-generator/issues/10) ChatGPT integration to automatically generate JSDoc descriptions.
+- New contribution `generateJsdocFolder` to generate JSDocs for all files in a folder, available with the contextual menu in the Explorer view.
 
 ### Fixed